follower forecast functionality
Created a new entity>followers_hist.py, to handle the method calls for the follower history in firebase.

Created a folder 'generate_fake_datas', to store the files used for generating fake informations in firebase

Added the overview for followers gained this month in the influencer dashboard.

Some other changes just to make it neater (not important)